requirements.txt
anyio==4.6.0
certifi==2024.8.30
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pillow==10.4.0
pyaes==1.6.1
pyasn1==0.6.1
python-telegram-bot==21.6
rsa==4.9
sniffio==1.3.1
Telethon==1.37.0
typing_extensions==4.12.2
